+++ packages/open/src/ShortcutHandler.tsx
@@ -1,0 +1,51 @@
+/**
+ * A React component to view a PDF document
+ *
+ * @see https://react-pdf-viewer.dev
+ * @license https://react-pdf-viewer.dev/license
+ * @copyright 2019-2021 Nguyen Huu Phuoc <me@phuoc.ng>
+ */
+
+import * as React from 'react';
+import { isMac } from '@react-pdf-viewer/core';
+import type { Store } from '@react-pdf-viewer/core';
+
+import type { StoreProps } from './types/StoreProps';
+
+export const ShortcutHandler: React.FC<{
+    containerRef: React.RefObject<HTMLDivElement>;
+    store: Store<StoreProps>;
+}> = ({ containerRef, store }) => {
+    const keydownHandler = (e: KeyboardEvent) => {
+        if (e.shiftKey || e.altKey || e.key !== 'o') {
+            return;
+        }
+
+        const isCommandPressed = isMac() ? e.metaKey : e.ctrlKey;
+        if (!isCommandPressed) {
+            return;
+        }
+
+        const containerEle = containerRef.current;
+        if (!containerEle || !document.activeElement || !containerEle.contains(document.activeElement)) {
+            return;
+        }
+
+        e.preventDefault();
+        store.update('triggerOpenFile', true);
+    };
+
+    React.useEffect(() => {
+        const containerEle = containerRef.current;
+        if (!containerEle) {
+            return;
+        }
+
+        document.addEventListener('keydown', keydownHandler);
+        return () => {
+            document.removeEventListener('keydown', keydownHandler);
+        };
+    }, [containerRef.current]);
+
+    return <></>;
+};