+++ packages/open/src/openPlugin.tsx
@@ -8,10 +8,11 @@
 
 import * as React from 'react';
 import { createStore } from '@react-pdf-viewer/core';
-import type { Plugin, PluginFunctions } from '@react-pdf-viewer/core';
+import type { Plugin, PluginFunctions, RenderViewer, Slot } from '@react-pdf-viewer/core';
 
 import { Open, OpenProps } from './Open';
 import { OpenMenuItem } from './OpenMenuItem';
+import { ShortcutHandler } from './ShortcutHandler';
 import type { StoreProps } from './types/StoreProps';
 
 export interface OpenPlugin extends Plugin {
@@ -20,19 +21,42 @@ export interface OpenPlugin extends Plugin {
     OpenMenuItem: () => React.ReactElement;
 }
 
-export const openPlugin = (): OpenPlugin => {
+export interface OpenPluginProps {
+    enableShortcuts?: boolean;
+}
+
+export const openPlugin = (props?: OpenPluginProps): OpenPlugin => {
+    const openPluginProps = React.useMemo(() => Object.assign({}, { enableShortcuts: true }, props), []);
     const store = React.useMemo(() => createStore<StoreProps>({}), []);
 
     const OpenDecorator = (props: OpenProps) => <Open {...props} store={store} />;
 
     const OpenButtonDecorator = () => <OpenDecorator />;
 
-    const OpenMenuItemDecorator = () => <OpenDecorator>{(p) => <OpenMenuItem onClick={p.onClick} />}</OpenDecorator>;
+    const OpenMenuItemDecorator = () => (
+        <OpenDecorator>{(p) => <OpenMenuItem store={store} onClick={p.onClick} />}</OpenDecorator>
+    );
+
+    const renderViewer = (props: RenderViewer): Slot => {
+        const { slot } = props;
+        const updateSlot: Slot = {
+            children: (
+                <>
+                    {openPluginProps.enableShortcuts && (
+                        <ShortcutHandler containerRef={props.containerRef} store={store} />
+                    )}
+                    {slot.children}
+                </>
+            ),
+        };
+        return { ...slot, ...updateSlot };
+    };
 
     return {
         install: (pluginFunctions: PluginFunctions) => {
             store.update('openFile', pluginFunctions.openFile);
         },
+        renderViewer,
         Open: OpenDecorator,
         OpenButton: OpenButtonDecorator,
         OpenMenuItem: OpenMenuItemDecorator,