+++ packages/open/src/useTriggerOpen.ts
@@ -1,0 +1,45 @@
+/**
+ * A React component to view a PDF document
+ *
+ * @see https://react-pdf-viewer.dev
+ * @license https://react-pdf-viewer.dev/license
+ * @copyright 2019-2021 Nguyen Huu Phuoc <me@phuoc.ng>
+ */
+
+import * as React from 'react';
+import type { Store } from '@react-pdf-viewer/core';
+
+import type { StoreProps } from './types/StoreProps';
+
+export const useTriggerOpen = (store: Store<StoreProps>) => {
+    const inputRef = React.useRef<HTMLInputElement>();
+
+    const openFile = () => {
+        const inputEle = inputRef.current;
+        if (inputEle) {
+            inputEle.click();
+            if (store.get('triggerOpenFile')) {
+                store.update('triggerOpenFile', false);
+            }
+        }
+    };
+
+    const handleOpenFileTriggered = (trigger: boolean) => {
+        if (trigger) {
+            openFile();
+        }
+    };
+
+    React.useEffect(() => {
+        store.subscribe('triggerOpenFile', handleOpenFileTriggered);
+
+        return () => {
+            store.unsubscribe('triggerOpenFile', handleOpenFileTriggered);
+        };
+    }, []);
+
+    return {
+        inputRef,
+        openFile,
+    };
+};